# mnt/private/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API de développement"
        }
    ],
    "paths": {
        "/api/abilities": {
            "get": {
                "tags": [
                    "Abilities"
                ],
                "summary": "Get a list of all abilities",
                "operationId": "8878dd9cc0d17f160a9e7b01390e12ff",
                "responses": {
                    "200": {
                        "description": "List of abilities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ability"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Abilities"
                ],
                "summary": "Create a new ability",
                "operationId": "641e36f6e546c4ce41884c5932246049",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "SKILL_ID",
                                    "ABI_LABEL"
                                ],
                                "properties": {
                                    "SKILL_ID": {
                                        "type": "integer"
                                    },
                                    "ABI_LABEL": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ability created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ability"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/abilities/{id}": {
            "get": {
                "tags": [
                    "Abilities"
                ],
                "summary": "Get a specific ability by ID",
                "operationId": "f563e02d738fdc9404c02b777e85675e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ability",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ability found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ability"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ability not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Abilities"
                ],
                "summary": "Update an existing ability",
                "operationId": "4c20730d4bfd162b4b789fea3caaae57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ability",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ability"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ability updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ability"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Ability not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Abilities"
                ],
                "summary": "Delete an ability",
                "operationId": "3d2395d936cd1464c9f44d7b5834e041",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ability",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ability deleted successfully"
                    },
                    "404": {
                        "description": "Ability not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ability": {
                "required": [
                    "id",
                    "SKILL_ID",
                    "ABI_LABEL"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the ability",
                        "type": "integer"
                    },
                    "SKILL_ID": {
                        "description": "Skill ID associated with the ability",
                        "type": "integer"
                    },
                    "ABI_LABEL": {
                        "description": "Label of the ability",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}